+++ hammurabi/rules/abstract.py
@@ -21,9 +21,16 @@ class AbstractRule(ABC):
     :type param: Any
     """
 
+    __slots__ = ("param", "name", "made_changes")
+
     def __init__(self, name: str, param: Any) -> None:
         self.param = param
         self.name = name.strip()
+
+        # Set by GitMixin or other mixins to indicate that the rule did changes.
+        # Rules can set this flag directly too. Only those rules will be indicated on
+        # Git commit which are made changes.
+        self.made_changes = False
 
     @staticmethod
     def validate(val: Any, cast_to: Optional[Any] = None, required=False) -> Any: