+++ tests/helpers.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from hammurabi.mixins import GitHubMixin, GitMixin
+from hammurabi.mixins import GitHubMixin, GitMixin, PullRequestHelperMixin
 from hammurabi.rules.base import Precondition, Rule
 
 
@@ -12,17 +12,29 @@ class ExamplePrecondition(Precondition):
 class ExampleRule(Rule):
     """ExampleRule docstring"""
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.made_changes = True
+
     def task(self) -> Any:
         """ExampleRule task docstring"""
         return self.param
 
 
 class ExampleExceptionRule(Rule):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.made_changes = False
+
     def task(self) -> Any:
         raise Exception(self.param)
 
 
 class ExampleGitMixinRule(ExampleRule, GitMixin):
+    pass
+
+
+class ExamplePullRequestHelperMixinRule(ExampleRule, GitMixin, PullRequestHelperMixin):
     pass
 
 
@@ -34,8 +46,8 @@ PASSING_PRECONDITION = ExamplePrecondition(name="Passing", param=True)
 FAILING_PRECONDITION = ExamplePrecondition(name="Failing", param=False)
 
 
-def get_passing_rule():
-    return ExampleRule(name="Passing", param="passing rule")
+def get_passing_rule(name: str = "Passing"):
+    return ExampleRule(name=name, param="passing rule")
 
 
 def get_failing_rule():
@@ -44,6 +56,10 @@ def get_failing_rule():
 
 def get_git_mixin_consumer():
     return ExampleGitMixinRule(name="Passing", param="passing rule")
+
+
+def get_pull_request_helper_mixin_consumer():
+    return ExamplePullRequestHelperMixinRule(name="Passing", param="passing rule")
 
 
 def get_github_mixin_consumer():